% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelL1.R
\name{LinearModelL1}
\alias{LinearModelL1}
\title{Linear Model with L1 regularization}
\usage{
LinearModelL1(x.scaled.mat, y.vec, penalty, opt.thresh, initial.weight.vec,
  step.size = 0.5)
}
\arguments{
\item{x.scaled.mat}{matrix of size [n x p]}

\item{y.vec}{vector of size n}

\item{penalty}{non negative numeric scalar}

\item{opt.thresh}{positive numeric scalar}

\item{initial.weight.vec}{vector of initial weights}

\item{step.size}{integer > 1}
}
\value{
optimal.weight.vector
}
\description{
Linear Model with L1 regularization
}
