% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelL1Penalties.R
\name{LinearModelL1penalties}
\alias{LinearModelL1penalties}
\title{Linear Model with L1 regularization}
\usage{
LinearModelL1penalties(x.mat, y.vec, penalty.vec, step.size)
}
\arguments{
\item{x.mat}{matrix of size [n x p]}

\item{y.vec}{vector of size n}

\item{penalty.vec}{vector of decreasing penalty values}

\item{step.size}{a step size for each step}
}
\value{
optimal.weight.vector
}
\description{
This algorithm takes the penalty vector as input and iterate through each value in the vector using LinearModelL1
}
